note_common: add file for note* functions
These note* functions are being used by both issue_ and mr_ commands, so moving
it to a common file seems better than keep it in a issue_ file not being
restricted to that.
