+++ cmd/issue_note_test.go
@@ -5,7 +5,6 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
@@ -22,36 +21,6 @@ func Test_issueCreateNote(t *testing.T) {
 	}
 
 	require.Contains(t, string(b), "https://gitlab.com/lab-testing/test/issues/1#note_")
-}
-
-func Test_issueNoteMsg(t *testing.T) {
-	tests := []struct {
-		Name         string
-		Msgs         []string
-		ExpectedBody string
-	}{
-		{
-			Name:         "Using messages",
-			Msgs:         []string{"note paragraph 1", "note paragraph 2"},
-			ExpectedBody: "note paragraph 1\n\nnote paragraph 2",
-		},
-		{
-			Name:         "From Editor",
-			Msgs:         nil,
-			ExpectedBody: "", // this is not a great test
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.Name, func(t *testing.T) {
-			test := test
-			t.Parallel()
-			body, err := noteMsg(test.Msgs, false, "\n")
-			if err != nil {
-				t.Fatal(err)
-			}
-			assert.Equal(t, test.ExpectedBody, body)
-		})
-	}
 }
 
 func Test_issueReplyNote(t *testing.T) {
@@ -104,16 +73,4 @@ func Test_issueReplyNote(t *testing.T) {
 	require.Contains(t, string(d), "#"+replyID+": "+"lab-testing commented at")
 	require.Contains(t, string(d), "    reply to note")
 	require.Contains(t, string(d), "    second reply paragraph")
-}
-
-func Test_issueNoteText(t *testing.T) {
-	t.Parallel()
-	text, err := noteText("\n")
-	if err != nil {
-		t.Fatal(err)
-	}
-	require.Equal(t, `
-
-# Write a message for this note. Commented lines are discarded.`, text)
-
 }
